Added 5 python files and deleted 1 rock_paper_scissor file
endless_run.py --> function for infinite rounds
round_of_mode1.py --> function for round of gaming mode
rules_of_game.py --> function to display the rules of the game
final_results.py --> function to display the results of the game after
playing any of the gaming mode
main_menu.py --> This is the file which controls the whole game.
All the above files are imported in the main file and are linked with
one another through this file.

To execute the game on your local computer you have to execute the
main_menu.py file